requirements.txt
python-dotenv
schedule
pandas
python-binance
supabase
twilio